+++ piGantry/piSerial/__init__.py
@@ -1,3 +1,4 @@
 # __init__.py
 from .serialInter import *
 from .TkInterUI import *
+from .motorFunc import *